+++ .vscode/launch.json
@@ -4,6 +4,7 @@
     // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
     "version": "0.2.0",
     "configurations": [
+        {"name":"Python Debugger: Current File","type":"debugpy","request":"launch","program":"${file}","console":"integratedTerminal"},
         {
             "name": "Python Debugger: Flask",
             "type": "debugpy",
@@ -35,6 +36,6 @@
           }
         ]
       }
-      
+
     ]
 }
